Move typeid cli to typeid/ repo (#24)
Move the typeid cli to the `typeid` repo, instead of `typeid-go`. This
way, `typeid-go` will be solely for the go library. The top level
`typeid` repo provides a cli tool to generates `typids` from the
command-line. It happens to be written in go, but that's incidental.

This is a tiny PR that I wrote using Devbox Cloud to go through the
whole experience.