Change how linting works in opensource (#440)
## Community Contribution License
All community contributions in this pull request are licensed to the
project maintainers
under the terms of the [Apache 2
license](https://www.apache.org/licenses/LICENSE-2.0).
By creating this pull request I represent that I have the right to
license the
contributions to the project maintainers under the Apache 2 license.

## Summary
Change how linting works in opensource to get it passing again. As of
now, the version from devbox cannot work with go 1.24, so I switched to
using their official github action extension.

## How was it tested?
This PR will show me if CICD passes or not.