Proposal for v2 of typeid (#502)
## Summary
Implements a proposed version 2 of TypeID:
+ Removes all the generic stuff and simplifies the API considerably. 
+ Improves the base32 encoder to allow for zero allocations and buffer
reuse.
+ Improved error handling via CorruptInputError and ErrValidation.
+ Adds extensive benchmarks. Rewrote existing ones as well. All
benchmarks now use the newer loop notation and use sinks (to prevent
compiler optimizations from invalidating the benchmarks)
+ Rethinks the SQL integration to have a nullable version and a
non-nullable version.
+ Changes the internal encoding of the TypeID struct to be more
efficient by using a single string and an integer. Note that I tested a
byte slice version as well, and it didn't perform as well as this one in
the benchmarks.
+ Improved unit tests so that every single constructor is tested against
the spec.
+ Rewrote constructors to minimize allocations as much as possible (some
constructors are zero allocations, others are 1 or 2 at most)
+ Changed meaning of IsZero() (so that it makes sense to use with
`omitzero`). Introduced HasSuffix() that addresses the same use case as
the previous version of IsZero().

Rollout proposal:
+ Merge this PR
+ Update axiom to use it, see if we like it better. Maybe we tweak it a
bit more.
+ Once happy with it, we release a new version under `v2`

## How was it tested?
Ran all the tests

## Community Contribution License

All community contributions in this pull request are licensed to the
project maintainers under the terms of the [Apache 2
License](https://www.apache.org/licenses/LICENSE-2.0).

By creating this pull request I represent that I have the right to
license the contributions to the project maintainers under the Apache 2
License as stated in the [Community Contribution
License](https://github.com/jetify-com/opensource/blob/main/CONTRIBUTING.md#community-contribution-license).